--- a/Jint/Key.cs
+++ b/Jint/Key.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
+using Jint.Native;
 using Jint.Runtime;
 
 namespace Jint
@@ -23,22 +24,42 @@ namespace Jint
             }
         }
 
-        public Key(string name)
+        public Key(string name) : this(name, symbolIdentity: 0)
+        {
+        }
+
+        internal Key(string name, int symbolIdentity)
         {
             if (name == null)
             {
                 ExceptionHelper.ThrowArgumentException("name cannot be null");
             }
             Name = name;
-            HashCode = name.GetHashCode();
+            HashCode = name.GetHashCode() * 397 ^ symbolIdentity.GetHashCode();
+            _symbolIdentity = symbolIdentity;
         }
 
         public readonly string Name;
         internal readonly int HashCode;
+        internal readonly int _symbolIdentity;
+
+        public Types Type => _symbolIdentity == 0 ? Types.String : Types.Symbol;
+
+        public bool IsSymbol => _symbolIdentity != 0;
 
         public static implicit operator Key(string name)
         {
             return new Key(name);
+        }
+
+        public static implicit operator Key(JsSymbol name)
+        {
+            return name.ToPropertyKey();
+        }
+
+        public static implicit operator JsValue(in Key key)
+        {
+            return !key.IsSymbol ? (JsValue) JsString.Create(key.Name) : new JsSymbol(key.Name, key._symbolIdentity);
         }
 
         public static implicit operator Key(int value)
@@ -53,6 +74,12 @@ namespace Jint
             return value < keys.Length ? keys[value] : BuildKey(value);
         }
 
+        public static implicit operator Key(ulong value)
+        {
+            var keys = indexKeys;
+            return value < (ulong) keys.Length ? keys[value] : new Key(value.ToString());
+        }
+
         [MethodImpl(MethodImplOptions.NoInlining)]
         private static Key BuildKey(long value)
         {
@@ -63,27 +90,27 @@ namespace Jint
 
         public static bool operator ==(in Key a, Key b)
         {
-            return a.HashCode == b.HashCode && a.Name == b.Name;
+            return a.HashCode == b.HashCode && a.Name == b.Name && a._symbolIdentity == b._symbolIdentity;
         }
 
         public static bool operator !=(in Key a, Key b)
         {
-            return a.HashCode != b.HashCode || a.Name != b.Name;
+            return a.HashCode != b.HashCode || a.Name != b.Name || a._symbolIdentity != b._symbolIdentity;
         }
 
         public static bool operator ==(in Key a, string b)
         {
-            return a.Name == b;
+            return a.Name == b && a._symbolIdentity == 0;
         }
 
         public static bool operator !=(in Key a, string b)
         {
-            return a.Name != b;
+            return a.Name != b || a._symbolIdentity > 0;
         }
 
         public bool Equals(Key other)
         {
-            return HashCode == other.HashCode && Name == other.Name;
+            return Name == other.Name && _symbolIdentity == other._symbolIdentity;
         }
 
         public override bool Equals(object obj)