--- a/Jint/Native/IConstructor.cs
+++ b/Jint/Native/IConstructor.cs
@@ -4,7 +4,6 @@ namespace Jint.Native
 {
     public interface IConstructor
     {
-        JsValue Call(JsValue thisObject, JsValue[] arguments);
-        ObjectInstance Construct(JsValue[] arguments);
+        ObjectInstance Construct(JsValue[] arguments, JsValue newTarget);
     }
 }