OK this was quite the chore. I tried to align as much as I could with the spec and changed method names like Put to Set, which is what the spec now uses. Also named helper methods with spec names.

Key can now handle both symbols and "regular" properties, fixed some issues with identities too. Later we should probably rename Key to PropertyKey.

Generally I think some things got cleaned up, sorry for the large PR but the proxy and reflect walk quite hand in hand. Also fixed existing issues that were revealed when proxy support was enabled for tests that were in ignore before. Almost 400 new tests covering the new functionality.