@sebastienros 2019-12-26
Looks great, thank you. I see you updated BDN, but I don't see numbers, so I assume it's bad ;)