@lahma 2019-12-27
Here are the numbers, generally some logic got simplified, but don't want to deviate too much from spec for performance now that there are big things coming like let/const 😉 

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
AMD Ryzen 7 2700X, 1 CPU, 16 logical and 8 physical cores
.NET Core SDK=3.1.100
  [Host]     : .NET Core 3.0.1 (CoreCLR 4.700.19.51502, CoreFX 4.700.19.51609), X64 RyuJIT
  DefaultJob : .NET Core 3.0.1 (CoreCLR 4.700.19.51502, CoreFX 4.700.19.51609), X64 RyuJIT


```

## Esprima.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Run|3d-cube|411.4 ms|4.28 ms|11000.0000|1000.0000|0.0000|49 MB|
| **New** |	|	| **394.9 ms (-4%)** | **3.32 ms** | **11000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **49.27 MB (+1%)** |
| Old |Run|3d-morph|345.7 ms|2.01 ms|10000.0000|2000.0000|0.0000|65.85 MB|
| **New** |	|	| **356.4 ms (+3%)** | **2.04 ms** | **10000.0000 (0%)** | **2000.0000 (0%)** | **0.0000** | **67.17 MB (+2%)** |
| Old |Run|3d-raytrace|310.3 ms|1.68 ms|20000.0000|3000.0000|0.0000|84.42 MB|
| **New** |	|	| **315.8 ms (+2%)** | **0.72 ms** | **20000.0000 (0%)** | **3000.0000 (0%)** | **0.0000** | **84.96 MB (+1%)** |
| Old |Run|access-binary-trees|151.8 ms|0.42 ms|12500.0000|500.0000|0.0000|52.19 MB|
| **New** |	|	| **155.4 ms (+2%)** | **0.55 ms** | **13500.0000 (+8%)** | **500.0000 (0%)** | **0.0000** | **56.08 MB (+7%)** |
| Old |Run|access-fannkuch|1,026.3 ms|13.25 ms|9000.0000|0.0000|0.0000|37.91 MB|
| **New** |	|	| **1,023.1 ms (0%)** | **9.39 ms** | **9000.0000 (0%)** | **0.0000** | **0.0000** | **37.91 MB (0%)** |
| Old |Run|access-nbody|363.7 ms|1.03 ms|13000.0000|0.0000|0.0000|52.18 MB|
| **New** |	|	| **362.2 ms (0%)** | **6.99 ms** | **15000.0000 (+15%)** | **0.0000** | **0.0000** | **62.49 MB (+20%)** |
| Old |Run|access-nsieve|405.2 ms|1.30 ms|11000.0000|3000.0000|0.0000|61.71 MB|
| **New** |	|	| **401.0 ms (-1%)** | **3.09 ms** | **11000.0000 (0%)** | **3000.0000 (0%)** | **0.0000** | **61.71 MB (0%)** |
| Old |Run|bitop(...)-byte [24]|247.3 ms|1.37 ms|12000.0000|0.0000|0.0000|51.73 MB|
| **New** |	|	| **234.8 ms (-5%)** | **1.32 ms** | **12000.0000 (0%)** | **0.0000** | **0.0000** | **51.73 MB (0%)** |
| Old |Run|bitops-bits-in-byte|401.8 ms|2.75 ms|8000.0000|0.0000|0.0000|33.46 MB|
| **New** |	|	| **376.7 ms (-6%)** | **2.92 ms** | **8000.0000 (0%)** | **0.0000** | **0.0000** | **33.46 MB (0%)** |
| Old |Run|bitops-bitwise-and|193.3 ms|0.90 ms|4333.3333|0.0000|0.0000|18.01 MB|
| **New** |	|	| **299.4 ms (+55%)** | **2.36 ms** | **13000.0000 (+200%)** | **0.0000** | **0.0000** | **54.64 MB (+203%)** |
| Old |Run|bitops-nsieve-bits|383.2 ms|2.03 ms|19000.0000|1000.0000|0.0000|80.19 MB|
| **New** |	|	| **377.6 ms (-1%)** | **1.80 ms** | **19000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **80.49 MB (0%)** |
| Old |Run|contr(...)rsive [21]|209.8 ms|0.84 ms|14000.0000|2000.0000|0.0000|56.4 MB|
| **New** |	|	| **216.4 ms (+3%)** | **1.13 ms** | **14000.0000 (0%)** | **2000.0000 (0%)** | **0.0000** | **56.4 MB (0%)** |
| Old |Run|crypto-aes|277.6 ms|0.88 ms|5000.0000|1000.0000|0.0000|22.3 MB|
| **New** |	|	| **270.8 ms (-2%)** | **0.85 ms** | **5000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **22.31 MB (0%)** |
| Old |Run|crypto-md5|199.0 ms|0.83 ms|19000.0000|1000.0000|0.0000|78.74 MB|
| **New** |	|	| **201.8 ms (+1%)** | **1.22 ms** | **19000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **78.74 MB (0%)** |
| Old |Run|crypto-sha1|190.9 ms|1.24 ms|12000.0000|2000.0000|0.0000|51.13 MB|
| **New** |	|	| **193.1 ms (+1%)** | **1.94 ms** | **12000.0000 (0%)** | **2000.0000 (0%)** | **0.0000** | **51.13 MB (0%)** |
| Old |Run|date-format-tofte|233.8 ms|0.34 ms|15000.0000|1000.0000|0.0000|60.63 MB|
| **New** |	|	| **230.9 ms (-1%)** | **1.06 ms** | **15000.0000 (0%)** | **0.0000 (-100%)** | **0.0000** | **60.68 MB (0%)** |
| Old |Run|date-format-xparb|250.4 ms|0.40 ms|4000.0000|1000.0000|0.0000|19.36 MB|
| **New** |	|	| **246.0 ms (-2%)** | **0.69 ms** | **4000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **19.73 MB (+2%)** |
| Old |Run|math-cordic|594.6 ms|5.28 ms|14000.0000|0.0000|0.0000|56.82 MB|
| **New** |	|	| **583.6 ms (-2%)** | **1.71 ms** | **14000.0000 (0%)** | **0.0000** | **0.0000** | **57.58 MB (+1%)** |
| Old |Run|math-partial-sums|198.7 ms|1.90 ms|6000.0000|0.0000|0.0000|27.63 MB|
| **New** |	|	| **238.3 ms (+20%)** | **2.86 ms** | **9000.0000 (+50%)** | **0.0000** | **0.0000** | **38.29 MB (+39%)** |
| Old |Run|math-spectral-norm|203.5 ms|1.46 ms|6000.0000|0.0000|0.0000|27.47 MB|
| **New** |	|	| **196.9 ms (-3%)** | **0.97 ms** | **6000.0000 (0%)** | **0.0000** | **0.0000** | **27.47 MB (0%)** |
| Old |Run|regexp-dna|335.2 ms|0.55 ms|0.0000|0.0000|0.0000|19.36 MB|
| **New** |	|	| **350.2 ms (+4%)** | **0.70 ms** | **0.0000** | **0.0000** | **0.0000** | **19.37 MB (0%)** |
| Old |Run|string-base64|165.4 ms|0.97 ms|8000.0000|1000.0000|0.0000|34.56 MB|
| **New** |	|	| **165.7 ms (0%)** | **0.80 ms** | **8000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **35.06 MB (+1%)** |
| Old |Run|string-fasta|313.6 ms|2.67 ms|22000.0000|0.0000|0.0000|90.09 MB|
| **New** |	|	| **327.0 ms (+4%)** | **2.37 ms** | **22000.0000 (0%)** | **0.0000** | **0.0000** | **89.17 MB (-1%)** |
| Old |Run|string-tagcloud|148.9 ms|0.97 ms|9000.0000|1000.0000|0.0000|49.73 MB|
| **New** |	|	| **150.8 ms (+1%)** | **3.00 ms** | **9000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **49.58 MB (0%)** |
| Old |Run|string-unpack-code|140.8 ms|1.34 ms|7000.0000|1000.0000|0.0000|40.77 MB|
| **New** |	|	| **138.4 ms (-2%)** | **0.89 ms** | **8000.0000 (+14%)** | **1000.0000 (0%)** | **0.0000** | **41.07 MB (+1%)** |
| Old |Run|strin(...)input [21]|130.8 ms|0.73 ms|6000.0000|1000.0000|0.0000|25.44 MB|
| **New** |	|	| **131.2 ms (0%)** | **0.59 ms** | **6000.0000 (0%)** | **1000.0000 (0%)** | **0.0000** | **26.11 MB (+3%)** |


## Jint.Benchmark.ArrayBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Slice|100|358.4 us|1.19 us|171.8750|0.0000|0.0000|703.91 KB|
| **New** |	|	| **374.3 us (+4%)** | **1.78 us** | **171.8750 (0%)** | **0.0000** | **0.0000** | **703.91 KB (0%)** |
| Old |Concat|100|487.0 us|2.67 us|187.5000|0.0000|0.0000|766.41 KB|
| **New** |	|	| **501.5 us (+3%)** | **3.13 us** | **187.5000 (0%)** | **0.0000** | **0.0000** | **766.41 KB (0%)** |
| Old |Unshift|100|15,703.5 us|32.03 us|3625.0000|0.0000|0.0000|14815.63 KB|
| **New** |	|	| **16,251.8 us (+3%)** | **42.57 us** | **3687.5000 (+2%)** | **0.0000** | **0.0000** | **15078.13 KB (+2%)** |
| Old |Push|100|6,850.2 us|45.55 us|390.6250|0.0000|0.0000|1605.5 KB|
| **New** |	|	| **7,964.7 us (+16%)** | **25.34 us** | **453.1250 (+16%)** | **0.0000** | **0.0000** | **1867.99 KB (+16%)** |
| Old |Index|100|7,952.4 us|24.39 us|437.5000|0.0000|0.0000|1829.69 KB|
| **New** |	|	| **9,237.5 us (+16%)** | **71.35 us** | **500.0000 (+14%)** | **0.0000** | **0.0000** | **2092.19 KB (+14%)** |
| Old |Map|100|3,157.6 us|11.43 us|261.7188|0.0000|0.0000|1083.6 KB|
| **New** |	|	| **2,985.3 us (-5%)** | **19.73 us** | **265.6250 (+1%)** | **0.0000** | **0.0000** | **1086.72 KB (0%)** |
| Old |Apply|100|557.1 us|1.62 us|202.1484|0.0000|0.0000|828.13 KB|
| **New** |	|	| **574.1 us (+3%)** | **7.80 us** | **202.1484 (0%)** | **0.0000** | **0.0000** | **828.13 KB (0%)** |
| Old |JsonStringifyParse|100|4,181.8 us|10.97 us|1226.5625|0.0000|0.0000|5038.29 KB|
| **New** |	|	| **4,392.3 us (+5%)** | **13.05 us** | **1226.5625 (0%)** | **0.0000** | **0.0000** | **5038.29 KB (0%)** |
| Old |FilterWithString|100|33,528.1 us|502.70 us|133.3333|0.0000|0.0000|770.31 KB|
| **New** |	|	| **31,581.3 us (-6%)** | **191.72 us** | **187.5000 (+41%)** | **0.0000** | **0.0000** | **770.31 KB (0%)** |


## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Run|dromaeo-3d-cube|48.99 ms|0.106 ms|1363.6364|272.7273|0.0000|7806.56 KB|
| **New** |	|	| **49.73 ms (+2%)** | **0.195 ms** | **1400.0000 (+3%)** | **400.0000 (+47%)** | **0.0000** | **7845.09 KB (0%)** |
| Old |Run|dromaeo-core-eval|11.77 ms|0.091 ms|78.1250|0.0000|0.0000|376.61 KB|
| **New** |	|	| **11.82 ms (0%)** | **0.126 ms** | **78.1250 (0%)** | **15.6250** | **0.0000** | **377.83 KB (0%)** |
| Old |Run|dromaeo-object-array|120.83 ms|0.184 ms|41200.0000|1400.0000|400.0000|173426.84 KB|
| **New** |	|	| **125.09 ms (+4%)** | **0.304 ms** | **42000.0000 (+2%)** | **1750.0000 (+25%)** | **250.0000 (-38%)** | **175092.86 KB (+1%)** |
| Old |Run|droma(...)egexp [21]|628.76 ms|10.143 ms|49000.0000|28000.0000|19000.0000|409938.09 KB|
| **New** |	|	| **644.24 ms (+2%)** | **11.694 ms** | **51000.0000 (+4%)** | **32000.0000 (+14%)** | **20000.0000 (+5%)** | **409301.94 KB (0%)** |
| Old |Run|droma(...)tring [21]|470.62 ms|8.850 ms|112000.0000|72000.0000|70000.0000|1371576.39 KB|
| **New** |	|	| **495.92 ms (+5%)** | **4.666 ms** | **115000.0000 (+3%)** | **73000.0000 (+1%)** | **71000.0000 (+1%)** | **1378369.21 KB (0%)** |


## Jint.Benchmark.UncacheableExpressionsBenchmark

| **Diff**|Method|N|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |Benchmark|500|209.7 ms|1.11 ms|18333.3333|333.3333|0.0000|73.93 MB|
| **New** |	|	| **213.2 ms (+2%)** | **0.89 ms** | **18333.3333 (0%)** | **333.3333 (0%)** | **0.0000** | **73.93 MB (0%)** |